+++ changelog.md
@@ -23,6 +23,7 @@
   - Arrow Signs
   - Large Arrow Signs
 - Added Arrow Signs for all wood types, with Every Compat support
+- Added Hard Wooden Blocks for all wood types, with Every Compat support
 - Added tooltips to bridges
 - Wandering trader trades Mushroot Saplings
 - Fixed flammability of Log Stairs